+++ module/handlebars/templates.js
@@ -34,6 +34,7 @@ export const preloadHandlebarsTemplates = async function () {
     "systems/pf1/templates/items/parts/item-contents.hbs",
     "systems/pf1/templates/items/parts/item-tag.hbs",
     "systems/pf1/templates/items/parts/item-name.hbs",
+    "systems/pf1/templates/items/parts/item-advanced.hbs",
 
     // Apps
     "systems/pf1/templates/apps/attack-roll-dialog.hbs",